backend/CloudRun-Base/requirements.txt
requests
fastapi==0.109.2
uvicorn==0.27.1
natsort==8.4.0
python-dotenv==1.0.1
lxml==5.1.0
pandas==2.2.2
polars==0.20.17
